+++ features/support/hooks.rb
@@ -1,5 +1,6 @@
 Before do
   VCR.use_cassette("clean_database_for_testing") do
-    RestClient.post "http://#{WEBSERVER}:#{HTTP_PORT_NUMBER}/SampleService/RubyOData.svc/CleanDatabaseForTesting", {}
+    conn = Faraday.new(url: "http://#{WEBSERVER}:#{HTTP_PORT_NUMBER}")
+    conn.post '/SampleService/RubyOData.svc/CleanDatabaseForTesting'
   end
 end