+++ spec/revised_service_spec.rb
@@ -1,4 +1,5 @@
 require 'spec_helper'
+require 'base64'
 
 module OData
 
@@ -172,7 +173,7 @@ module OData
         end
         it "should return true if a function import post that returns successfully and doesn't have a return value (HTTP 204)" do
           result = subject.CleanDatabaseForTesting
-          result.should be_true
+          expect(result).to eq true
         end
         it "should return a collection of entities for a collection" do
           result = subject.EntityCategoryWebGet
@@ -219,7 +220,7 @@ module OData
 
         its(:name)                { should eq('id') }
         its(:type)                { should eq('Edm.Int64') }
-        its(:nullable)            { should be_false }
+        its(:nullable)            { should eq false }
         its(:fc_target_path)      { should be_nil }
         its(:fc_keep_in_content)  { should be_nil }
       end
@@ -260,7 +261,7 @@ module OData
 
         its(:name)                { should eq('KeyId') }
         its(:type)                { should eq('Edm.Int64') }
-        its(:nullable)            { should be_false }
+        its(:nullable)            { should eq(false) }
         its(:fc_target_path)      { should be_nil }
         its(:fc_keep_in_content)  { should be_nil }
       end
@@ -286,10 +287,13 @@ module OData
 
   describe "Dual Namespaces" do
     before(:all) do
-      stub_request(:get, "http://xxxx%5Cyyyy:zzzz@test.com/test.svc/$metadata").
-        with(:headers => DEFAULT_HEADERS).
+      auth_string = "xxxx\\yyyy:zzzz"
+      authorization_header = { authorization: "Basic #{Base64::encode64(auth_string).strip}" }
+      stub_request(:get, "http://test.com/test.svc/$metadata").
+        with(:headers => DEFAULT_HEADERS.merge(authorization_header)).
         to_return(:status => 200, :body => File.new(File.expand_path("../fixtures/ms_system_center/edmx_ms_system_center_v2.xml", __FILE__)), :headers => {})
     end
+
     after(:all) do
       VMM.constants.each do |constant|
         VMM.send :remove_const, constant